pi-agent/requirements.txt
pyaudio>=0.2.13
PyYAML>=6.0
requests>=2.31.0
numpy>=1.24.0
